requirements.txt
flask==3.0.2
flask-wtf==1.2.1
openai==0.28.1
python-dotenv==1.0.1
nltk==3.8.1
python-docx==1.1.0
werkzeug==3.0.1
gunicorn==21.2.0
PyPDF2==3.0.1 